Removed Community modularization from Gumby
